requirements.txt
Django==1.9.1
Pillow==3.1.0
PyPDF2==1.25.1
django-filter==0.12.0
djangorestframework==3.3.2
djangular==0.2.7
html5lib==0.9999999
importlib==1.0.3
reportlab==3.2.0
six==1.10.0
xhtml2pdf==0.0.6
